requirement.txt
Django==1.8.6
MySQL-python==1.2.5
django-mysql==1.0.0
djangorestframework==3.3.2
django-oauth-toolkit==0.10.0
oauth2-provider
django-cors-headers==1.1.0
boto3==1.2.3

